[ADD] vendor_bills: Submodule added as dependency
I'm adding this submodule here because we depends on
`l10n_mx_customer_bills`.

When we use t2d and our traditional containers this repository is
automatically fetched by the script `clone_oca_dependencies` because it
is a dependency of the `mexico` repository
[here](https://git.vauxoo.com/vauxoo/mexico/blob/5cf06832dc84d1713e24d9897e86b56fe8e4816e/oca_dependencies.txt#L3).
But it is not fetched automatically by Odoo.sh. So I need to put it
here explicitly.