[FIX] amper: Fix duplicated value on AMLs for stock valuation
When a purchase order is made for the same product twice but with
different price, the same value is set on the stock AMLs after
validating the vendor bill. That is because the action to update AMLs
when a vendor bill is validated worked looking for the product.

To solve the above, The action was re-written to take into account not
only the product, but the price in USD.

A migration script is provided to fix the purchase order `PO00249`,
which is currently the only affected by this issue.